requirements.txt
# Core dependencies
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
asyncio>=3.4.3

# API and networking
urllib3>=2.0.0
certifi>=2023.0.0
websockets>=11.0.0
aiohttp>=3.8.0

# AI and intelligence
openai>=1.0.0
python-dotenv>=1.0.0

# Data analysis and visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Utilities
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Enhanced dependencies
ta-lib>=0.4.0
ccxt>=4.0.0
websocket-client>=1.6.0

# On-chain analytics and enhanced API clients
httpx>=0.25.0  # For async HTTP requests (Perplexity, Dune)
tqdm>=4.66.0   # Progress bars for CLI
rich>=13.0.0   # Enhanced CLI output
# dune-client>=0.7.0  # Uncomment if using official Dune client (optional)

# New dependencies
scipy>=1.10.0
tensorflow>=2.12.0
scikit-learn>=1.2.0
joblib>=1.2.0
